requirements.txt
Django==5.2.7
djangorestframework==3.15.2
graphene-django==3.1.5
graphql-core==3.2.3
requests==2.31.0

